cpplint: fix the majority of runtime/int
Bug: 364967694
Test: m .
Flag: Exempt refactor

+++ system/gd/common/strings.h
@@ -53,22 +53,21 @@ std::string ToHexString(T x) {
     return "-" + ToHexString(-x);
   }
   std::stringstream tmp;
-  tmp << "0x" << std::internal << std::hex << std::setfill('0') << std::setw(sizeof(T) * 2)
-      << (unsigned long)x;
+  tmp << "0x" << std::internal << std::hex << std::setfill('0') << std::setw(sizeof(T) * 2) << x;
   return tmp.str();
 }
 
 template <>
-inline std::string ToHexString<>(signed long x) {
+inline std::string ToHexString<>(int64_t x) {
   if (x < 0) {
     if (x == LONG_MIN) {
       return "LONG_MIN";
     }
-    return "-" + ToHexString<signed long>(-x);
+    return "-" + ToHexString<int64_t>(-x);
   }
   std::stringstream tmp;
-  tmp << "0x" << std::internal << std::hex << std::setfill('0')
-      << std::setw(sizeof(signed long) * 2) << (unsigned long)x;
+  tmp << "0x" << std::internal << std::hex << std::setfill('0') << std::setw(sizeof(int64_t) * 2)
+      << x;
   return tmp.str();
 }
 
@@ -76,7 +75,7 @@ template <>
 inline std::string ToHexString<>(unsigned int x) {
   std::stringstream tmp;
   tmp << "0x" << std::internal << std::hex << std::setfill('0')
-      << std::setw(sizeof(unsigned int) * 2) << (unsigned long)x;
+      << std::setw(sizeof(unsigned int) * 2) << x;
   return tmp.str();
 }
